--- app/src/main/java/com/seunghoon/generator/di/KoinApplication.kt
@@ -1,11 +1,0 @@
-package com.seunghoon.generator.di
-
-import android.app.Application
-import org.koin.core.context.startKoin
-
-class KoinApplication : Application() {
-    override fun onCreate() {
-        super.onCreate()
-        startKoin {}
-    }
-}
