--- app/src/main/java/com/seunghoon/generator/navigation/ProjectGeneratorApp.kt
@@ -1,18 +1,0 @@
-package com.seunghoon.generator.navigation
-
-import androidx.compose.runtime.Composable
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.rememberNavController
-
-@Composable
-internal fun ProjectGeneratorApp() {
-    val navController = rememberNavController()
-
-    NavHost(
-        navController = navController,
-        startDestination = NavigationRoute.Auth.route,
-    ) {
-        auth()
-        main()
-    }
-}
