+++ app/src/main/java/gram/killergram/app/di/KoinApplication.kt
@@ -1,0 +1,11 @@
+package gram.killergram.app.di
+
+import android.app.Application
+import org.koin.core.context.startKoin
+
+class KoinApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        startKoin {}
+    }
+}
