+++ app/src/main/java/gram/killergram/app/navigation/ProjectGeneratorApp.kt
@@ -1,0 +1,18 @@
+package gram.killergram.app.navigation
+
+import androidx.compose.runtime.Composable
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.rememberNavController
+
+@Composable
+internal fun ProjectGeneratorApp() {
+    val navController = rememberNavController()
+
+    NavHost(
+        navController = navController,
+        startDestination = NavigationRoute.Auth.route,
+    ) {
+        auth()
+        main()
+    }
+}
