--- core/designsystem/src/main/java/com/seunghoon/designsystem/ui/theme/Theme.kt
@@ -1,46 +1,0 @@
-package com.seunghoon.designsystem.ui.theme
-
-import android.app.Activity
-import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.material3.darkColorScheme
-import androidx.compose.material3.lightColorScheme
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.SideEffect
-import androidx.compose.runtime.staticCompositionLocalOf
-import androidx.compose.ui.graphics.toArgb
-import androidx.compose.ui.platform.LocalView
-import androidx.core.view.WindowCompat
-
-private val DarkColorScheme = darkColorScheme()
-
-private val LightColorScheme = lightColorScheme()
-
-@Composable
-fun ProjectGeneratorTheme(
-    darkTheme: Boolean = isSystemInDarkTheme(),
-    content: @Composable () -> Unit,
-) {
-    val colorScheme = when {
-        darkTheme -> DarkColorScheme
-        else -> LightColorScheme
-    }
-    val view = LocalView.current
-    if (!view.isInEditMode) {
-        SideEffect {
-            val window = (view.context as Activity).window
-            window.statusBarColor = colorScheme.primary.toArgb()
-            WindowCompat.getInsetsController(window, view).isAppearanceLightStatusBars = darkTheme
-        }
-    }
-
-    CompositionLocalProvider(LocalColors provides colorScheme) {
-        content()
-    }
-}
-
-// 이름 변경해서 사용
-val Theme
-    @Composable get() = LocalColors.current
-
-private val LocalColors = staticCompositionLocalOf { lightColorScheme() }
