--- core/designsystem/src/main/java/com/seunghoon/designsystem/utils/KeyboardAsState.kt
@@ -1,14 +1,0 @@
-package com.seunghoon.designsystem.utils
-
-import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.ime
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.State
-import androidx.compose.runtime.rememberUpdatedState
-import androidx.compose.ui.platform.LocalDensity
-
-@Composable
-fun keyboardAsState(): State<Boolean> {
-    val isImeVisible = WindowInsets.ime.getBottom(LocalDensity.current) > 0
-    return rememberUpdatedState(newValue = isImeVisible)
-}
