+++ core/designsystem/src/main/java/gram/killergram/core/designsystem/ui/icon/GramIcons.kt
@@ -1,0 +1,22 @@
+package gram.killergram.core.designsystem.ui.icon
+
+import com.seunghoon.core.designsystem.R
+
+object GramIcons {
+    val BADMINTON = R.drawable.ic_badminton
+    val VOLLEYBALL = R.drawable.ic_volleyball
+    val BASKETBALL = R.drawable.ic_basketball
+    val WOMAN = R.drawable.ic_woman
+    val TABLE_TENNIS = R.drawable.ic_table_tennis
+    val WEIGHT_TRAINING = R.drawable.ic_weight_training
+    val BASEBALL = R.drawable.ic_baseball
+    val SOCCER = R.drawable.ic_soccer
+    val ADD = R.drawable.ic_add
+    val ARROW_DOWN = R.drawable.ic_arrow_down
+    val ARROW_LEFT = R.drawable.ic_arrow_left
+    val ARROW_RIGHT = R.drawable.ic_arrow_right
+    val ARROW_UP = R.drawable.ic_arrow_up
+    val CLOSE = R.drawable.ic_close
+    val OPTION = R.drawable.ic_option
+    val REFRESH = R.drawable.ic_refresh
+}
