+++ core/designsystem/src/main/java/gram/killergram/core/designsystem/ui/theme/GramColors.kt
@@ -1,0 +1,28 @@
+package gram.killergram.core.designsystem.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+object GramColors {
+    val White = Color(0xFFFFFFFF)
+    val Black = Color(0xFF333333)
+    val Transparent = Color(0x00000000)
+
+    val Main = Color(0xFF9EFF00)
+    val Secondary = Color(0xFF325100)
+
+    val Gray100 = Color(0xFFF5F6F8)
+    val Gray200 = Color(0xFFEFF0F2)
+    val Gray300 = Color(0xFFDBDCDE)
+    val Gray400 = Color(0xFFB4B5B7)
+    val Gray500 = Color(0xFFA5A6A9)
+    val Gray600 = Color(0xFF8F9094)
+    val Gray700 = Color(0xFF828387)
+    val Gray800 = Color(0xFF666669)
+    val Gray900 = Color(0xFF4F4F51)
+    val Gray1000 = Color(0xFF212121)
+    val Gray1100 = Color(0xFF181818)
+
+    val System = Color(0xFFE53A3A)
+
+    val Bakc = Color(0xFF0F0F0F)
+}
