+++ core/designsystem/src/main/java/gram/killergram/core/designsystem/utils/KeyboardAsState.kt
@@ -1,0 +1,14 @@
+package gram.killergram.core.designsystem.utils
+
+import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.ime
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.State
+import androidx.compose.runtime.rememberUpdatedState
+import androidx.compose.ui.platform.LocalDensity
+
+@Composable
+fun keyboardAsState(): State<Boolean> {
+    val isImeVisible = WindowInsets.ime.getBottom(LocalDensity.current) > 0
+    return rememberUpdatedState(newValue = isImeVisible)
+}
